Additional Code Review Changes.
