+++ test/Performance/project.json
@@ -5,7 +5,7 @@
       "type": "platform",
       "version": "1.0.0-rc2-*"
     },
-    "System.Runtime.Serialization.Primitives": "4.1.1-rc2-24015",
+    "System.Runtime.Serialization.Primitives": "4.1.1-rc2-24022",
     "Microsoft.DotNet.ProjectModel": {
       "target": "project"
     },
@@ -17,15 +17,14 @@
     },
     "xunit": "2.1.0",
     "xunit.netcore.extensions": "1.0.0-prerelease-00206",
-    "dotnet-test-xunit": "1.0.0-dev-140469-38",
+    "dotnet-test-xunit": "1.0.0-rc2-162081-13",
     "Microsoft.DotNet.xunit.performance": "1.0.0-alpha-build0028"
   },
   "frameworks": {
     "netcoreapp1.0": {
       "imports": [
-        "netstandardapp1.5",
-        "dnxcore50",
-        "portable-net45+win8"
+        "dotnet5.4",
+        "portable-net451+win8"
       ]
     }
   },
